JWT-service/requirements.txt
fastapi
uvicorn
motor
passlib[bcrypt]
PyJWT